+++ src/main.rs
@@ -7,9 +7,10 @@ pub mod util;
 use crate::{
     channels::{read_courses, MiEI},
     commands::{
-        admin::{ADMIN_GROUP, DAEMONS_GROUP, MINECRAFT_GROUP},
+        admin::ADMIN_GROUP,
         cesium::{ChannelMapping, CESIUM_GROUP},
-        COURSES_GROUP, MISC_GROUP, STUDY_GROUP,
+        misc::MISC_GROUP,
+        study::{COURSES_GROUP, STUDY_GROUP},
     },
     config::Config,
     daemons::minecraft::Minecraft,
@@ -187,8 +188,6 @@ fn main() {
             .group(&ADMIN_GROUP)
             .group(&MISC_GROUP)
             .group(&CESIUM_GROUP)
-            .group(&MINECRAFT_GROUP)
-            .group(&DAEMONS_GROUP)
             .help(&MY_HELP),
     );
     if let Err(why) = client.start() {
